kinopark-backend/requirements.txt
asgiref==3.10.0
attrs==25.4.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
django-unfold==0.67.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==12.0.0
PyJWT==2.10.1
python-decouple==3.8
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
